api/requirements.txt
fastapi
motor
google-genai
pydantic
dotenv
weasyprint
uvicorn
pymongo